requirements.txt
fastapi
uvicorn[standard]
python-binance
python-dotenv
jinja2